3scale: Eliminate custom container
These are expensive, as we would need to publish and maintain it. The
config files are better passed in as volumes (possibly even ConfigMaps,
or Secrets on production).

While at it, also avoid recursive Makefiles, and move everything into
the top-level one. Introduce a new `build` default target which will
build the credentials and our custom container image.