appservice: Use podman 3 compatible network option
podman < 4 does not yet understand the "Networks" property when creating
a container, and silently ignores it. This led to session pods being
attached to the default network, so the appservice pod could not reach
them.

Use the deprecated `cni_networks` option for the time being, it still
works with podman 4.