appservice: Fix nginx startup
Replace the sleep(1) hack with a proper waiting loop. We need to wait
for nginx to start up sufficiently to be able to properly react to
SIGHUP -- otherwise, the redis client could HUP it in the early startup
phase, and kill it.

Also initialize the NGINX_PID variable to None, so that we get a proper
crash if we try to signal it before instantiation. Otherwise, we'd kill
ourselves (pid 0 means "our process group"). Rename it to NGINX_PROC, so
that we can do operations on the Popen object.

Fixes #24