appservice: Fix for running as non-root
On kubernetes we don't get to run as root, but as some random system
user. To work in that environment, make the nginx config, state, and log
directories world-writable, and run nginx on port 8080 instead of the
priviled port 80.

Note that we can't use a non-root container user with a local podman deployment
(by e.g. setting `USER www-data`), as it would then not be permitted to access
the bind-mounted host podman socket. On kubernetes we don't have such a thing,
but talk to the k8s REST API, so it works there.