Rework port configuration/API URL
Specifying just the port is podman specific. For k8s we will have to
specify the full API URL, as that depends on 3scale and isn't
predictable from within the service. Thus change the multiplexer to expect
`$API_URL` to be set (which will then work for both podman and k8s).

Change the test to read the env variable from the deployed container, instead
of getting it passed in directly. This make the tests slightly more universal
and less dependent on being run from `make test`.