appservice: More robust forwarding of control paths
Only send known control paths to multiplexer.py. Otherwise this is going
to receive unknown or stale session paths as well, which looks weird in
the logs and isn't going to work. Let them fall through the 404 rule
instead.

Also forward them to the IPv4 localhost address only. With "localhost",
nginx sends them to ::1 as well. As multiplexer is not listening on
IPv6, this causes a lot of error messages like

    [error] 37#37: *31 connect() failed (111: Connection refused) while connecting to upstream
    "GET /api/webconsole/v1/sessions/new HTTP/1.0", upstream: "http://[::1]:8081/api/webconsole/v1/sessions/new"

which makes the multiplexer slower and debugging harder.