appservice: Rewrite using starlette and httpx
aiohttp has a long-standing race condition with reading replies [1]. This is
completely killing cockpit on kubernetes, and so far there is no workaround.

Thus move to a different framework. Starlette [1] is quite pleasant and
performant, and supports async and websockets. This also allows us to export
the API to consoledot as OpenAPI.

Many thanks to Christian Heimes @tiran for his initial code!

[1] https://github.com/aio-libs/aiohttp/issues/4581
[2] https://www.starlette.io/