Replace websocat on target machine with Python script
We want to minimize the requirements on the target machines. websocat
is a great tool, but not packaged anywhere, and downloading straight
from GitHub isn't appropriate for production.

Thus replace this with a simple Python connector which uses the builtin
asyncio module, and the external websockets one. To avoid having to
install the latter, build a pyz application bundle. We can then also
test/CI this before shipping.

The connector only supports basic auth for now. Eventually it will need
to do TLS client cert auth, but we'll get to that when we have the
necessary test harness.

Stop putting the connector into the webconsoleserver image. Eventually
it will be sent to the target machine through an Ansible script. Thus
for testing, send it to the container as a volume instead, which from
the point of the connector amounts to the same thing.

Fixes #41