appservice: Handle websocket close exceptions
These actually appear as `websockets.exceptions.ConnectionClosedOK`
exception instead of a particular message type, at least when logging
out or switching to a different URL and thus letting the cockpit
websocket time out.

Intercept these and orderly exit `ws_down2up`, so that websocket_forward() can
close the other ends of its websockets. Also update the session status to
"closed", so that we can handle this in the upcoming session stub page.