appservice: Cache session pod IP addresses
Resolve the session pod IP *once* on creation, instead of on every
proxied HTTP and websocket request. This speeds up proxying and also
makes the cause for errors more obvious. Tests previously often failed
with a long nested backtrace, whose initial reason was a
`socket.gaierror` when trying to resolve the DNS name in
handle_session_id_*().

This also fixes a race condition with k8s -- the DNS name is not
available immediately after starting a pod (as that is asynchronous), so
trying to proxy to the session right after creation failed. Add a retry
loop to the session creation to wait until the session pod becomes
resolvable in DNS, in other words, when it actually is running. This
also provides a more sensible error handling if starting the pod fails.