server: Fix cockpit-bridge-websocket-connector RHEL 8 compatibility
We want this script to work on RHEL/CentOS 8 as well, which only has
Python 3.6.

Unfortunately there is no websockets version that works for all Python versions
from 3.6 up to 3.12, so we need to install multiple versions. Install the most
recent version (8.0.x series at the moment) that is still compatible with
Python 3.6, rename it (and its internal imports) to `websockets36` to avoid
collision with the current version, and conditionally import it in the connector.

In the pip install command, `--no-deps` is mandatory with `--python-version`,
but that is okay -- websockets does not have any required dependencies (and our
tests will confirm that).

In the connector itself, provide some shims for APIs which don't yet
exist in 3.6.

Fixes #53