test: Ensure that bridge connector picks up EOF
When the session pod dies or the Cockpit session ends, the bridge
connector ought to pick up the EOF, close its end, and exit cleanly.
Ensure that this happens by validating that the webconsoleserver
container exists after crashing the session pod or timing out the
cockpit session.

Give the server pods a predictable name for that. This also makes it
easier to investigate failures, as session and server pods are now
nicely grouped, instead of the latter having auto-generated random
names.

Fixes #55