appservice: Clean up session container when session closes
As of version 276, cockpit-ws does not exit once the `--local-session`
script ends, nor has an option to do so. We want the session container
to exit itself after the session closes, so that it stops and gets
cleaned up. To do that, make websocket-session.py kill its parent
process (i.e. cockpit-ws) at the end.

For this to have any effect, cockpit-ws must *not* run as pid 1 in the
container, as that does not have any SIGKILL or similar default signal
handlers set up, and it does not listen to SIGTERM either. Thus keep the
original shell from cockpit-ws-session.sh around and start cockpit-ws in
a new process.

Note: An alternative implementation would be to explicitly kill the
container with the podman/kubernetes API. But our "deployer" service
account on k8s is not permitted to remove pods, and we can't (nor really
want to) create a custom service account role ourselves. Also, it is
more elegant to let session containers decide for themselves when it's
time to RIP.

Fixes #30