Add authentication and authorization
- parse 'x-rh-identity' header in authentication middleware
- set auth credential scope based on header presence and values

The authentication middle returns AuthCredentials and a user object. The
user object contains information about 'who' performed the request.
AuthCredentials define the scope for authorization. In our case the user
object also has the org id, type (`User` or `System`), and additional
information. For requests with a correct `x-rh-identity` header the
scope is `authenticated` and either `User` or `System`.

Starlette's authentication API mandates that a route function either
accepts a `request` argument for HTTP routes or a `websocket` argument
for WebSocket routes. Other argument names are not supported.